+++ CHANGES.rst
@@ -12,7 +12,7 @@ development
 2017-10-20 0.2.0
 ================
 - Python3 compatibility
-- Add commandline program "pairclient" for printing and saving documents
+- Add commandline program "uspto-pbd" for printing and saving documents from the USPTO PAIR Bulk Data API
 
 2017-10-20 0.1.0
 ================